feat: :heavy_plus_sign: Add lib/news.ts
Issue: #140
